pyproject.toml
[tool.poetry]
name = "to-do-list-tg-lab"
version = "0.1.0"
description = "A simple app to manage tasks."
authors = ["volodymyr-v-konoval <volodymyr.v.konoval@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
